fix: apply suggestions
