+++ electron.vite.config.ts
@@ -1,44 +1,17 @@
-import { existsSync } from 'node:fs';
-import { basename, relative, resolve } from 'node:path';
-
 import { defineConfig, defineViteConfig } from 'electron-vite';
 import builtinModules from 'builtin-modules';
 import viteResolve from 'vite-plugin-resolve';
 
-import { globSync } from 'glob';
+import { pluginVirtualModuleGenerator } from './vite-plugins/plugin-virtual-module-generator';
 
 import type { UserConfig } from 'vite';
-
-const pluginImportGenerator = (mode: 'back' | 'preload' | 'front') => {
-  let homePath = resolve(__dirname, 'src');
-
-  const plugins = globSync(`${homePath}/../src/plugins/*`)
-    .map((path) => ({ name: basename(path), path }))
-    .filter(({ name, path }) => !name.startsWith('utils') && existsSync(resolve(path, `${mode}.ts`)));
-
-  let result = '';
-
-  for (const { name, path } of plugins) {
-    result += `import ${snakeToCamel(name)}Plugin from "./${relative(resolve(homePath, '..'), path).replace(/\\/g, '/')}/${mode}";\n`;
-  }
-
-  result += `export const pluginList = {\n`;
-  for (const { name } of plugins) {
-    result += `  "${name}": ${snakeToCamel(name)}Plugin,\n`;
-  }
-  result += '};';
-
-  return result;
-}
-
-const snakeToCamel = (text: string) => text.replace(/-(\w)/g, (_, letter: string) => letter.toUpperCase());
 
 export default defineConfig({
   main: defineViteConfig(({ mode }) => {
     const commonConfig: UserConfig = {
       plugins: [
         viteResolve({
-          mainPlugins: pluginImportGenerator('back'),
+          'virtual:MainPlugins': pluginVirtualModuleGenerator('back'),
         }),
       ],
       publicDir: 'assets',
@@ -75,7 +48,7 @@ export default defineConfig({
     const commonConfig: UserConfig = {
       plugins: [
         viteResolve({
-          preloadPlugins: pluginImportGenerator('preload'),
+          'virtual:PreloadPlugins': pluginVirtualModuleGenerator('preload'),
         }),
       ],
       build: {
@@ -111,7 +84,7 @@ export default defineConfig({
     const commonConfig: UserConfig = {
       plugins: [
         viteResolve({
-          rendererPlugins: pluginImportGenerator('front'),
+          'virtual:RendererPlugins': pluginVirtualModuleGenerator('front'),
         }),
       ],
       root: './src/',
